fix(unstable-period): #144 — bound both sentinels on every surface
TA_Set/GetUnstablePeriod compared the id as a signed enum, so
TA_FUNC_UNST_NONE (-1) reached unstablePeriod[-1] — an out-of-bounds
write landing on the adjacent TA_Globals->compatibility, which then read
back as an invalid enumerator. One unsigned compare covers both ends and
stays correct whether the compiler picks a signed or unsigned underlying
type; -1 keeps its value, since the wrappers that pass it (ta-lib-python
and friends) hardcode it rather than read the header.

Rust's get_unstable_period(FuncUnstAll) indexed one past the array, as
filed; it now asserts, and CoreBuilder::candle_setting no longer
silently ignores AllCandleSettings — C returns TA_BAD_PARAM and Java
throws for that call. Java's CoreBuilder.unstablePeriod gained the
matching check on its write side.

The Java JSON-RPC server sized unstablePeriod by the enum's length and
compared the set-all sentinel against that length, so the driver's id 24
landed in an unused slot and "set all" was inert. Its enum now carries
both sentinels in the shipped order. A new cross-language gate proves
the wildcard reaches every function on all four servers, probing through
ADOSC (no unstablePeriod request field of its own, lookback follows
EMA's) — sabotage-proven against the old server.

Also: the FuncUnstId drift guard was inert at its generate-servers call
site (an input-relative root made it read a path that never exists), and
nothing ran the shipped Rust crate's unit tests — clippy --all-targets
only builds the test target and --doc skips it.

Closes #144.

+++ src/ta_func/ta_utility.c
@@ -45,6 +45,7 @@
  *  052603 MF     Adapt code to compile with .NET Managed C++
  *  123004 RM,MF  Adapt code to work with Visual Studio 2005
  *  071926 MF,CC  Remove dead .NET/Java preprocessor branches (plain C only)
+ *  072726 MF,CC  Bound TA_Set/GetUnstablePeriod below as well as above (#144)
  *
  */
 
@@ -57,7 +58,14 @@ TA_RetCode TA_SetUnstablePeriod( TA_FuncUnstId id,
 {
    int i;
 
-   if( id > TA_FUNC_UNST_ALL )
+   /* TA_FUNC_UNST_NONE (-1) forces a signed underlying type for the enum, so a
+    * signed-only "id > ALL" lets every negative id through to
+    * unstablePeriod[id] -- an out-of-bounds write onto the adjacent
+    * TA_Globals->compatibility. Comparing as unsigned wraps negatives past ALL,
+    * so one test covers both ends and stays correct whether the compiler picks
+    * a signed or unsigned type (any width up to unsigned int).
+    */
+   if( (unsigned int)id > (unsigned int)TA_FUNC_UNST_ALL )
       return TA_BAD_PARAM;
 
    if( id == TA_FUNC_UNST_ALL )
@@ -77,7 +85,10 @@ TA_RetCode TA_SetUnstablePeriod( TA_FuncUnstId id,
 
 unsigned int TA_GetUnstablePeriod( TA_FuncUnstId id )
 {
-   if( id >= TA_FUNC_UNST_ALL )
+   /* Unsigned compare -- see TA_SetUnstablePeriod above. Both sentinels and any
+    * out-of-range id read as 0 rather than off the end of the array.
+    */
+   if( (unsigned int)id >= (unsigned int)TA_FUNC_UNST_ALL )
 	   return 0;
 
    return TA_Globals->unstablePeriod[id];
